requirements.txt
selenium==4.15.2
selenium-wire==5.1.0
boto3==1.34.15
python-dotenv==1.1.0
requests==2.32.3
pika==1.3.2
webdriver-manager==4.0.1
blinker==1.7.
chromedriver-autoinstaller==0.6.4
fastapi==0.115.6
uvicorn==0.34.0
python-multipart==0.0.20
torch
open_clip_torch==2.23.0
pillow==10.2.0
numpy==1.24.4